PP-28: Правки после ревью. В модель Promocode добавлено поле is_active. ParentTokenObtainSerializer и ParentLoginAPIView заменены на UserTokenObtainSerialzier и UserLoginAPIView
