+++ backend/apps/user/api/v1/parent/serializers.py
@@ -11,30 +11,6 @@ from apps.user.models import User
 from apps.user.models.user import UserRole
 
 
-class ParentTokenObtainPairSerializer(TokenObtainPairSerializer):
-    """Сериализатор для авторизации родителя."""
-    def validate(self, attrs):
-        authenticate_kwargs = {
-            'email': attrs['email'],
-            'password': attrs['password'],
-        }
-        user = authenticate(**authenticate_kwargs)
-
-        if user is None or not user.is_active:
-            raise ValidationError('Нет такого пользователя.')
-
-        refresh = RefreshToken.for_user(user)
-
-        refresh['id'] = user.id
-        refresh['email'] = user.email
-        refresh['role'] = user.role
-
-        return {
-            'refresh': str(refresh),
-            'access': str(refresh.access_token),
-        }
-
-
 class EmailSerializer(serializers.Serializer):
     """Сериализатор для проверки email."""
     email = serializers.EmailField()
