+++ backend/apps/user/api/v1/serializers.py
@@ -1,7 +1,36 @@
 from rest_framework import serializers
+from rest_framework.exceptions import ValidationError
+from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
+from rest_framework_simplejwt.tokens import RefreshToken
+from django.contrib.auth import authenticate
+
 from django.contrib.auth.password_validation import validate_password
 
 from apps.user.models import User
+
+
+class UserTokenObtainPairSerializer(TokenObtainPairSerializer):
+    """Сериализатор для авторизации пользователя."""
+    def validate(self, attrs):
+        authenticate_kwargs = {
+            'email': attrs['email'],
+            'password': attrs['password'],
+        }
+        user = authenticate(**authenticate_kwargs)
+
+        if user is None or not user.is_active:
+            raise ValidationError('Нет такого пользователя.')
+
+        refresh = RefreshToken.for_user(user)
+
+        refresh['id'] = user.id
+        refresh['email'] = user.email
+        refresh['role'] = user.role
+
+        return {
+            'refresh': str(refresh),
+            'access': str(refresh.access_token),
+        }
 
 
 class UserSerializer(serializers.ModelSerializer):
